Early version of the review drawer
I'm still not sure what to do with that, either we
use AOSP behavior and scroll the right to the gallery and
use the review drawer simply to quickly review burst shots for
instance, or we fully use the review drawer for all the shots.
It needs a lot of refinement on all sides, but I need a playground
to test how I could make a proper use of that.

I also prepared it for a "Quick Review", like for examples
stacking shots from burst shot in a smaller version of the
review drawer, so user can see them directly and eventually tap
them right after being done with the burst.

There's some good ideas behind that, but I'll need some
experiment first.