add one-time camera check to disable Focal icon
Users of devices w/o cameras complain of the "useless" Focal in the launcher.

By copying the DisableCameraReceiver from Gallery2 into Focal, it now does a one-time check for cameras on the device (nearly the
same behaviour as the camera portion of Gallery2 -- see below). If none are found, all Focal activites are disabled (hiding the icon
in the launcher).

Slight difference in the disable check compared to the Gallery2 camera check: Focal's check only hides the icon if there are NO
cameras found, where the Gallery2 camera icon will be hidden if there are no REAR cameras found.

Change-Id: I0c9cef37acb99ea83fc74ee68c22ecca3cd34c14