[WIP] Focal: Revamp the sidebar UI
This commit revamps the way the side bar
and widgets UI works. While keeping it
mostly similar to how it used to work, it has
been changed to better fit quick gestures.

The sidebar now have bigger icons and a short
text description beneath each item (Note to
translators: Some text might be too big, there
is a limit of 2 lines to keep it clean).
The widget panel also uses bigger icons and are
scrollable, with the description of each icon
scrolling in marquee. A widget can be dismissed
by swiping its "title bar" to the right.
Widgets can also now be pinned to the main screen,
so that they are immediately accessible without
having to open the sidebar. Their toggle icon
then appears on the main screen instead of the
SideBar.

This also fixes the cling not showcasing the
views at the right place, and maybe one or two
other bugs.

This is still a work-in-progress and a few bugs
are to be expected (especially in landscape, I've
mostly tested it in portrait).

Change-Id: Iddac7f25d7db76bc93843b10068d4245fdce1e31