Optimize app startup and add Profiler class
Removed unneeded sounds that were preloaded,
as well as a few unneeded things.
Added a profiler class to log time used by certain
parts of the code.
Fixes a bug also where the app would enter
an infinite loop of NPE in CameraApplication
when the app crashes before mCamManager is
set.

Change-Id: If41b9c8cd0ae075084508d66ee4b4f50451cfd97